+++ src/Vu/ui/AllCommentPane.java
@@ -1,6 +1,6 @@
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/AWTForms/Dialog.java to edit this template
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
  */
 package Vu.ui;
 
@@ -8,13 +8,12 @@ package Vu.ui;
  *
  * @author Admin
  */
-public class AllCommentDialog extends java.awt.Dialog {
+public class AllCommentPane extends javax.swing.JPanel {
 
     /**
-     * Creates new form UserManagerment
+     * Creates new form AllCommentPane
      */
-    public AllCommentDialog(java.awt.Frame parent, boolean modal) {
-        super(parent, modal);
+    public AllCommentPane() {
         initComponents();
     }
 
@@ -23,12 +22,10 @@ public class AllCommentDialog extends java.awt.Dialog {
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
      */
+    @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
-        jPanel1 = new javax.swing.JPanel();
-        toolbar1 = new AdminToolUtils.Toolbar();
-        jPanel3 = new javax.swing.JPanel();
         jPanel2 = new javax.swing.JPanel();
         jPanel11 = new javax.swing.JPanel();
         findTextField1 = new AdminToolUtils.FindTextField();
@@ -44,31 +41,23 @@ public class AllCommentDialog extends java.awt.Dialog {
         jPanel6 = new javax.swing.JPanel();
         lblUser1 = new javax.swing.JLabel();
         lblComment1 = new javax.swing.JLabel();
-        btnDelete1 = new AdminToolUtils.Button();
+        btnDelete1 = new AdminToolUtils.ButtonDelete();
         jPanel7 = new javax.swing.JPanel();
         lblUser2 = new javax.swing.JLabel();
         lblComment2 = new javax.swing.JLabel();
-        btnDelete2 = new AdminToolUtils.Button();
+        btnDelete2 = new AdminToolUtils.ButtonDelete();
         jPanel8 = new javax.swing.JPanel();
         lblUser3 = new javax.swing.JLabel();
         lblComment3 = new javax.swing.JLabel();
-        btnDelete3 = new AdminToolUtils.Button();
+        btnDelete3 = new AdminToolUtils.ButtonDelete();
         jPanel9 = new javax.swing.JPanel();
         lblUser4 = new javax.swing.JLabel();
         lblComment4 = new javax.swing.JLabel();
-        btnDelete4 = new AdminToolUtils.Button();
+        btnDelete4 = new AdminToolUtils.ButtonDelete();
         jPanel10 = new javax.swing.JPanel();
         lblUser5 = new javax.swing.JLabel();
         lblComment5 = new javax.swing.JLabel();
-        btnDelete5 = new AdminToolUtils.Button();
-
-        addWindowListener(new java.awt.event.WindowAdapter() {
-            public void windowClosing(java.awt.event.WindowEvent evt) {
-                closeDialog(evt);
-            }
-        });
-
-        jPanel3.setLayout(new java.awt.CardLayout());
+        btnDelete5 = new AdminToolUtils.ButtonDelete();
 
         jPanel2.setBackground(new java.awt.Color(36, 36, 36));
 
@@ -420,82 +409,46 @@ public class AllCommentDialog extends java.awt.Dialog {
                 .addComponent(jPanel11, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, Short.MAX_VALUE))
-        );
-
-        jPanel3.add(jPanel2, "card2");
-
-        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
-        jPanel1.setLayout(jPanel1Layout);
-        jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addComponent(toolbar1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, 0)
-                .addComponent(jPanel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(toolbar1, javax.swing.GroupLayout.DEFAULT_SIZE, 807, Short.MAX_VALUE)
-            .addComponent(jPanel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addGap(0, 20, Short.MAX_VALUE))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
         this.setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addGap(0, 1220, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+            .addGap(0, 807, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
         );
-
-        pack();
     }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * Closes the dialog
-     */
-    private void closeDialog(java.awt.event.WindowEvent evt) {//GEN-FIRST:event_closeDialog
-        setVisible(false);
-        dispose();
-    }//GEN-LAST:event_closeDialog
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                AllCommentDialog dialog = new AllCommentDialog(new java.awt.Frame(), true);
-                dialog.addWindowListener(new java.awt.event.WindowAdapter() {
-                    public void windowClosing(java.awt.event.WindowEvent e) {
-                        System.exit(0);
-                    }
-                });
-                dialog.setVisible(true);
-            }
-        });
-    }
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
-    private AdminToolUtils.Button btnDelete1;
-    private AdminToolUtils.Button btnDelete2;
-    private AdminToolUtils.Button btnDelete3;
-    private AdminToolUtils.Button btnDelete4;
-    private AdminToolUtils.Button btnDelete5;
+    private AdminToolUtils.ButtonDelete btnDelete1;
+    private AdminToolUtils.ButtonDelete btnDelete2;
+    private AdminToolUtils.ButtonDelete btnDelete3;
+    private AdminToolUtils.ButtonDelete btnDelete4;
+    private AdminToolUtils.ButtonDelete btnDelete5;
     private AdminToolUtils.ComboBox comboBox2;
     private AdminToolUtils.FindTextField findTextField1;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
-    private javax.swing.JPanel jPanel1;
     private javax.swing.JPanel jPanel10;
     private javax.swing.JPanel jPanel11;
     private javax.swing.JPanel jPanel2;
-    private javax.swing.JPanel jPanel3;
     private javax.swing.JPanel jPanel4;
     private javax.swing.JPanel jPanel5;
     private javax.swing.JPanel jPanel6;
@@ -515,6 +468,5 @@ public class AllCommentDialog extends java.awt.Dialog {
     private javax.swing.JLabel lblUser4;
     private javax.swing.JLabel lblUser5;
     private javax.swing.JPanel pnlTitle;
-    private AdminToolUtils.Toolbar toolbar1;
     // End of variables declaration//GEN-END:variables
 }
