+++ src/Vu/ui/CensorSongPane.java
@@ -1,0 +1,695 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
+ */
+package Vu.ui;
+
+/**
+ *
+ * @author Admin
+ */
+public class CensorSongPane extends javax.swing.JPanel {
+
+    /**
+     * Creates new form CensorSongPane
+     */
+    public CensorSongPane() {
+        initComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jPanel1 = new javax.swing.JPanel();
+        jPanel2 = new javax.swing.JPanel();
+        findTextField2 = new AdminToolUtils.FindTextField();
+        cboSong = new AdminToolUtils.ComboBox();
+        jPanel4 = new javax.swing.JPanel();
+        pnlTitle = new javax.swing.JPanel();
+        lblNumber = new javax.swing.JLabel();
+        lblName = new javax.swing.JLabel();
+        lblUser = new javax.swing.JLabel();
+        lblStatus = new javax.swing.JLabel();
+        lblOption = new javax.swing.JLabel();
+        jSeparator2 = new javax.swing.JSeparator();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        jPanel5 = new javax.swing.JPanel();
+        jPanel6 = new javax.swing.JPanel();
+        lblNumber1 = new javax.swing.JLabel();
+        lblName1 = new javax.swing.JLabel();
+        lblUser1 = new javax.swing.JLabel();
+        lblUser9 = new javax.swing.JLabel();
+        btnDelete1 = new AdminToolUtils.ButtonDelete();
+        jPanel7 = new javax.swing.JPanel();
+        jPanel8 = new javax.swing.JPanel();
+        lblNumber2 = new javax.swing.JLabel();
+        lblName2 = new javax.swing.JLabel();
+        lblUser2 = new javax.swing.JLabel();
+        lblUser10 = new javax.swing.JLabel();
+        btnDelete2 = new AdminToolUtils.ButtonDelete();
+        jPanel9 = new javax.swing.JPanel();
+        lblNumber3 = new javax.swing.JLabel();
+        lblName3 = new javax.swing.JLabel();
+        lblUser3 = new javax.swing.JLabel();
+        lblUser11 = new javax.swing.JLabel();
+        btnDelete3 = new AdminToolUtils.ButtonDelete();
+        jPanel10 = new javax.swing.JPanel();
+        lblNumber4 = new javax.swing.JLabel();
+        lblName4 = new javax.swing.JLabel();
+        lblUser4 = new javax.swing.JLabel();
+        lblUser12 = new javax.swing.JLabel();
+        btnDelete4 = new AdminToolUtils.ButtonDelete();
+        jPanel11 = new javax.swing.JPanel();
+        lblNumber5 = new javax.swing.JLabel();
+        lblName5 = new javax.swing.JLabel();
+        lblUser5 = new javax.swing.JLabel();
+        lblUser13 = new javax.swing.JLabel();
+        btnDelete5 = new AdminToolUtils.ButtonDelete();
+        jPanel13 = new javax.swing.JPanel();
+        lblNumber7 = new javax.swing.JLabel();
+        lblName7 = new javax.swing.JLabel();
+        lblUser7 = new javax.swing.JLabel();
+        lblUser15 = new javax.swing.JLabel();
+        btnDelete7 = new AdminToolUtils.ButtonDelete();
+
+        setPreferredSize(new java.awt.Dimension(1220, 807));
+
+        jPanel1.setBackground(new java.awt.Color(36, 36, 36));
+
+        jPanel2.setBackground(new java.awt.Color(36, 36, 36));
+
+        findTextField2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        cboSong.setForeground(new java.awt.Color(0, 0, 0));
+        cboSong.setModel(new javax.swing.DefaultComboBoxModel(new String[] { "Tất cả bài hát", "Bài hát chưa kiểm duyệt" }));
+        cboSong.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
+        jPanel2.setLayout(jPanel2Layout);
+        jPanel2Layout.setHorizontalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel2Layout.createSequentialGroup()
+                .addGap(79, 79, 79)
+                .addComponent(findTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 450, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 204, Short.MAX_VALUE)
+                .addComponent(cboSong, javax.swing.GroupLayout.PREFERRED_SIZE, 400, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(87, 87, 87))
+        );
+        jPanel2Layout.setVerticalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
+                .addContainerGap(77, Short.MAX_VALUE)
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(findTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(cboSong, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(46, 46, 46))
+        );
+
+        jPanel4.setBackground(new java.awt.Color(36, 36, 36));
+
+        pnlTitle.setBackground(new java.awt.Color(36, 36, 36));
+        pnlTitle.setPreferredSize(new java.awt.Dimension(1073, 52));
+
+        lblNumber.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblNumber.setForeground(new java.awt.Color(199, 199, 199));
+        lblNumber.setText("STT");
+
+        lblName.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblName.setForeground(new java.awt.Color(199, 199, 199));
+        lblName.setText("Tên người dùng");
+
+        lblUser.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser.setForeground(new java.awt.Color(199, 199, 199));
+        lblUser.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser.setText("Nhạc sĩ");
+
+        lblStatus.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblStatus.setForeground(new java.awt.Color(199, 199, 199));
+        lblStatus.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblStatus.setText("Tùy chọn");
+
+        lblOption.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblOption.setForeground(new java.awt.Color(199, 199, 199));
+        lblOption.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblOption.setText("Ca sĩ");
+
+        javax.swing.GroupLayout pnlTitleLayout = new javax.swing.GroupLayout(pnlTitle);
+        pnlTitle.setLayout(pnlTitleLayout);
+        pnlTitleLayout.setHorizontalGroup(
+            pnlTitleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnlTitleLayout.createSequentialGroup()
+                .addGap(50, 50, 50)
+                .addComponent(lblNumber)
+                .addGap(64, 64, 64)
+                .addComponent(lblName)
+                .addGap(255, 255, 255)
+                .addComponent(lblUser)
+                .addGap(240, 240, 240)
+                .addComponent(lblOption)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(lblStatus)
+                .addGap(60, 60, 60))
+        );
+        pnlTitleLayout.setVerticalGroup(
+            pnlTitleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnlTitleLayout.createSequentialGroup()
+                .addGap(16, 16, 16)
+                .addGroup(pnlTitleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNumber)
+                    .addComponent(lblName)
+                    .addComponent(lblUser)
+                    .addComponent(lblStatus)
+                    .addComponent(lblOption))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+
+        jSeparator2.setBackground(new java.awt.Color(199, 199, 199));
+        jSeparator2.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+
+        jScrollPane1.setBackground(new java.awt.Color(0, 0, 0));
+        jScrollPane1.setBorder(null);
+        jScrollPane1.setHorizontalScrollBarPolicy(javax.swing.ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
+
+        jPanel5.setBackground(new java.awt.Color(36, 36, 36));
+
+        jPanel6.setBackground(new java.awt.Color(36, 36, 36));
+
+        lblNumber1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblNumber1.setForeground(new java.awt.Color(255, 255, 255));
+        lblNumber1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblNumber1.setText("1");
+
+        lblName1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblName1.setForeground(new java.awt.Color(255, 255, 255));
+        lblName1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
+        lblName1.setText("Mây hồng đưa lối");
+
+        lblUser1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser1.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser1.setText("Taylor Swift");
+
+        lblUser9.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser9.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser9.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser9.setText("Đạt Ka");
+
+        btnDelete1.setForeground(new java.awt.Color(0, 0, 0));
+        btnDelete1.setText("Xóa");
+        btnDelete1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        javax.swing.GroupLayout jPanel6Layout = new javax.swing.GroupLayout(jPanel6);
+        jPanel6.setLayout(jPanel6Layout);
+        jPanel6Layout.setHorizontalGroup(
+            jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel6Layout.createSequentialGroup()
+                .addGap(36, 36, 36)
+                .addComponent(lblNumber1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(55, 55, 55)
+                .addComponent(lblName1, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser1, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(100, 100, 100)
+                .addComponent(lblUser9, javax.swing.GroupLayout.PREFERRED_SIZE, 194, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 48, Short.MAX_VALUE)
+                .addComponent(btnDelete1, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(39, 39, 39))
+        );
+        jPanel6Layout.setVerticalGroup(
+            jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel6Layout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNumber1)
+                    .addComponent(lblName1)
+                    .addComponent(lblUser1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUser9, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnDelete1, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(21, Short.MAX_VALUE))
+        );
+
+        jPanel7.setBackground(new java.awt.Color(36, 36, 36));
+
+        jPanel8.setBackground(new java.awt.Color(36, 36, 36));
+
+        lblNumber2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblNumber2.setForeground(new java.awt.Color(255, 255, 255));
+        lblNumber2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblNumber2.setText("1");
+
+        lblName2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblName2.setForeground(new java.awt.Color(255, 255, 255));
+        lblName2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
+        lblName2.setText("Mây hồng đưa lối");
+
+        lblUser2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser2.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser2.setText("Taylor Swift");
+
+        lblUser10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser10.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser10.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser10.setText("Đạt Ka");
+
+        btnDelete2.setForeground(new java.awt.Color(0, 0, 0));
+        btnDelete2.setText("Xóa");
+        btnDelete2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        javax.swing.GroupLayout jPanel8Layout = new javax.swing.GroupLayout(jPanel8);
+        jPanel8.setLayout(jPanel8Layout);
+        jPanel8Layout.setHorizontalGroup(
+            jPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel8Layout.createSequentialGroup()
+                .addGap(36, 36, 36)
+                .addComponent(lblNumber2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(55, 55, 55)
+                .addComponent(lblName2, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser2, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(100, 100, 100)
+                .addComponent(lblUser10, javax.swing.GroupLayout.PREFERRED_SIZE, 194, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 48, Short.MAX_VALUE)
+                .addComponent(btnDelete2, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(39, 39, 39))
+        );
+        jPanel8Layout.setVerticalGroup(
+            jPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel8Layout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addGroup(jPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNumber2)
+                    .addComponent(lblName2)
+                    .addComponent(lblUser2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUser10, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnDelete2, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(33, Short.MAX_VALUE))
+        );
+
+        jPanel9.setBackground(new java.awt.Color(36, 36, 36));
+
+        lblNumber3.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblNumber3.setForeground(new java.awt.Color(255, 255, 255));
+        lblNumber3.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblNumber3.setText("1");
+
+        lblName3.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblName3.setForeground(new java.awt.Color(255, 255, 255));
+        lblName3.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
+        lblName3.setText("Mây hồng đưa lối");
+
+        lblUser3.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser3.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser3.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser3.setText("Taylor Swift");
+
+        lblUser11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser11.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser11.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser11.setText("Đạt Ka");
+
+        btnDelete3.setForeground(new java.awt.Color(0, 0, 0));
+        btnDelete3.setText("Xóa");
+        btnDelete3.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        javax.swing.GroupLayout jPanel9Layout = new javax.swing.GroupLayout(jPanel9);
+        jPanel9.setLayout(jPanel9Layout);
+        jPanel9Layout.setHorizontalGroup(
+            jPanel9Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel9Layout.createSequentialGroup()
+                .addGap(36, 36, 36)
+                .addComponent(lblNumber3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(55, 55, 55)
+                .addComponent(lblName3, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser3, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(100, 100, 100)
+                .addComponent(lblUser11, javax.swing.GroupLayout.PREFERRED_SIZE, 194, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 48, Short.MAX_VALUE)
+                .addComponent(btnDelete3, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(39, 39, 39))
+        );
+        jPanel9Layout.setVerticalGroup(
+            jPanel9Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel9Layout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addGroup(jPanel9Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNumber3)
+                    .addComponent(lblName3)
+                    .addComponent(lblUser3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUser11, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnDelete3, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(33, Short.MAX_VALUE))
+        );
+
+        jPanel10.setBackground(new java.awt.Color(36, 36, 36));
+
+        lblNumber4.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblNumber4.setForeground(new java.awt.Color(255, 255, 255));
+        lblNumber4.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblNumber4.setText("1");
+
+        lblName4.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblName4.setForeground(new java.awt.Color(255, 255, 255));
+        lblName4.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
+        lblName4.setText("Mây hồng đưa lối");
+
+        lblUser4.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser4.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser4.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser4.setText("Taylor Swift");
+
+        lblUser12.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser12.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser12.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser12.setText("Đạt Ka");
+
+        btnDelete4.setForeground(new java.awt.Color(0, 0, 0));
+        btnDelete4.setText("Xóa");
+        btnDelete4.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        javax.swing.GroupLayout jPanel10Layout = new javax.swing.GroupLayout(jPanel10);
+        jPanel10.setLayout(jPanel10Layout);
+        jPanel10Layout.setHorizontalGroup(
+            jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel10Layout.createSequentialGroup()
+                .addGap(36, 36, 36)
+                .addComponent(lblNumber4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(55, 55, 55)
+                .addComponent(lblName4, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser4, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(100, 100, 100)
+                .addComponent(lblUser12, javax.swing.GroupLayout.PREFERRED_SIZE, 194, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 48, Short.MAX_VALUE)
+                .addComponent(btnDelete4, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(39, 39, 39))
+        );
+        jPanel10Layout.setVerticalGroup(
+            jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel10Layout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNumber4)
+                    .addComponent(lblName4)
+                    .addComponent(lblUser4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUser12, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnDelete4, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(33, Short.MAX_VALUE))
+        );
+
+        jPanel11.setBackground(new java.awt.Color(36, 36, 36));
+
+        lblNumber5.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblNumber5.setForeground(new java.awt.Color(255, 255, 255));
+        lblNumber5.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblNumber5.setText("1");
+
+        lblName5.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblName5.setForeground(new java.awt.Color(255, 255, 255));
+        lblName5.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
+        lblName5.setText("Mây hồng đưa lối");
+
+        lblUser5.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser5.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser5.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser5.setText("Taylor Swift");
+
+        lblUser13.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser13.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser13.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser13.setText("Đạt Ka");
+
+        btnDelete5.setForeground(new java.awt.Color(0, 0, 0));
+        btnDelete5.setText("Xóa");
+        btnDelete5.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        javax.swing.GroupLayout jPanel11Layout = new javax.swing.GroupLayout(jPanel11);
+        jPanel11.setLayout(jPanel11Layout);
+        jPanel11Layout.setHorizontalGroup(
+            jPanel11Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel11Layout.createSequentialGroup()
+                .addGap(36, 36, 36)
+                .addComponent(lblNumber5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(55, 55, 55)
+                .addComponent(lblName5, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser5, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(100, 100, 100)
+                .addComponent(lblUser13, javax.swing.GroupLayout.PREFERRED_SIZE, 194, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 48, Short.MAX_VALUE)
+                .addComponent(btnDelete5, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(39, 39, 39))
+        );
+        jPanel11Layout.setVerticalGroup(
+            jPanel11Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel11Layout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addGroup(jPanel11Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNumber5)
+                    .addComponent(lblName5)
+                    .addComponent(lblUser5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUser13, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnDelete5, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(33, Short.MAX_VALUE))
+        );
+
+        jPanel13.setBackground(new java.awt.Color(36, 36, 36));
+
+        lblNumber7.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblNumber7.setForeground(new java.awt.Color(255, 255, 255));
+        lblNumber7.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblNumber7.setText("1");
+
+        lblName7.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblName7.setForeground(new java.awt.Color(255, 255, 255));
+        lblName7.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
+        lblName7.setText("Mây hồng đưa lối");
+
+        lblUser7.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser7.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser7.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser7.setText("Taylor Swift");
+
+        lblUser15.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser15.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser15.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser15.setText("Đạt Ka");
+
+        btnDelete7.setForeground(new java.awt.Color(0, 0, 0));
+        btnDelete7.setText("Xóa");
+        btnDelete7.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+
+        javax.swing.GroupLayout jPanel13Layout = new javax.swing.GroupLayout(jPanel13);
+        jPanel13.setLayout(jPanel13Layout);
+        jPanel13Layout.setHorizontalGroup(
+            jPanel13Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel13Layout.createSequentialGroup()
+                .addGap(36, 36, 36)
+                .addComponent(lblNumber7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(55, 55, 55)
+                .addComponent(lblName7, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser7, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(100, 100, 100)
+                .addComponent(lblUser15, javax.swing.GroupLayout.PREFERRED_SIZE, 194, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 48, Short.MAX_VALUE)
+                .addComponent(btnDelete7, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(39, 39, 39))
+        );
+        jPanel13Layout.setVerticalGroup(
+            jPanel13Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel13Layout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addGroup(jPanel13Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblNumber7)
+                    .addComponent(lblName7)
+                    .addComponent(lblUser7, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUser15, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnDelete7, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(21, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout jPanel7Layout = new javax.swing.GroupLayout(jPanel7);
+        jPanel7.setLayout(jPanel7Layout);
+        jPanel7Layout.setHorizontalGroup(
+            jPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel7Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jPanel8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jPanel9, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jPanel10, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jPanel11, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jPanel13, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        jPanel7Layout.setVerticalGroup(
+            jPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel7Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jPanel8, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPanel9, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPanel10, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPanel11, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPanel13, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(368, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout jPanel5Layout = new javax.swing.GroupLayout(jPanel5);
+        jPanel5.setLayout(jPanel5Layout);
+        jPanel5Layout.setHorizontalGroup(
+            jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel5Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jPanel6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addContainerGap())
+            .addGroup(jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(jPanel5Layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(jPanel7, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+        jPanel5Layout.setVerticalGroup(
+            jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel5Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jPanel6, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(804, Short.MAX_VALUE))
+            .addGroup(jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(jPanel5Layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(jPanel7, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+
+        jScrollPane1.setViewportView(jPanel5);
+
+        javax.swing.GroupLayout jPanel4Layout = new javax.swing.GroupLayout(jPanel4);
+        jPanel4.setLayout(jPanel4Layout);
+        jPanel4Layout.setHorizontalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel4Layout.createSequentialGroup()
+                .addGap(0, 0, 0)
+                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(jScrollPane1)
+                    .addComponent(jSeparator2)
+                    .addComponent(pnlTitle, javax.swing.GroupLayout.DEFAULT_SIZE, 1204, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        jPanel4Layout.setVerticalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel4Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(pnlTitle, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jSeparator2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 521, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+
+        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
+        jPanel1.setLayout(jPanel1Layout);
+        jPanel1Layout.setHorizontalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel1Layout.createSequentialGroup()
+                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, Short.MAX_VALUE))
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(45, 45, 45))
+        );
+        jPanel1Layout.setVerticalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel1Layout.createSequentialGroup()
+                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 22, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 1220, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 1220, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 807, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private AdminToolUtils.ButtonDelete btnDelete1;
+    private AdminToolUtils.ButtonDelete btnDelete2;
+    private AdminToolUtils.ButtonDelete btnDelete3;
+    private AdminToolUtils.ButtonDelete btnDelete4;
+    private AdminToolUtils.ButtonDelete btnDelete5;
+    private AdminToolUtils.ButtonDelete btnDelete7;
+    private AdminToolUtils.ComboBox cboSong;
+    private AdminToolUtils.FindTextField findTextField2;
+    private javax.swing.JPanel jPanel1;
+    private javax.swing.JPanel jPanel10;
+    private javax.swing.JPanel jPanel11;
+    private javax.swing.JPanel jPanel13;
+    private javax.swing.JPanel jPanel2;
+    private javax.swing.JPanel jPanel4;
+    private javax.swing.JPanel jPanel5;
+    private javax.swing.JPanel jPanel6;
+    private javax.swing.JPanel jPanel7;
+    private javax.swing.JPanel jPanel8;
+    private javax.swing.JPanel jPanel9;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JSeparator jSeparator2;
+    private javax.swing.JLabel lblName;
+    private javax.swing.JLabel lblName1;
+    private javax.swing.JLabel lblName2;
+    private javax.swing.JLabel lblName3;
+    private javax.swing.JLabel lblName4;
+    private javax.swing.JLabel lblName5;
+    private javax.swing.JLabel lblName7;
+    private javax.swing.JLabel lblNumber;
+    private javax.swing.JLabel lblNumber1;
+    private javax.swing.JLabel lblNumber2;
+    private javax.swing.JLabel lblNumber3;
+    private javax.swing.JLabel lblNumber4;
+    private javax.swing.JLabel lblNumber5;
+    private javax.swing.JLabel lblNumber7;
+    private javax.swing.JLabel lblOption;
+    private javax.swing.JLabel lblStatus;
+    private javax.swing.JLabel lblUser;
+    private javax.swing.JLabel lblUser1;
+    private javax.swing.JLabel lblUser10;
+    private javax.swing.JLabel lblUser11;
+    private javax.swing.JLabel lblUser12;
+    private javax.swing.JLabel lblUser13;
+    private javax.swing.JLabel lblUser15;
+    private javax.swing.JLabel lblUser2;
+    private javax.swing.JLabel lblUser3;
+    private javax.swing.JLabel lblUser4;
+    private javax.swing.JLabel lblUser5;
+    private javax.swing.JLabel lblUser7;
+    private javax.swing.JLabel lblUser9;
+    private javax.swing.JPanel pnlTitle;
+    // End of variables declaration//GEN-END:variables
+}
