+++ src/Vu/ui/Toolbar.form
@@ -48,7 +48,7 @@
               <Component id="btnStatistic" min="-2" pref="51" max="-2" attributes="0"/>
               <EmptySpace min="0" pref="0" max="-2" attributes="0"/>
               <Component id="btnExit" min="-2" pref="51" max="-2" attributes="0"/>
-              <EmptySpace pref="349" max="32767" attributes="0"/>
+              <EmptySpace pref="389" max="32767" attributes="0"/>
           </Group>
       </Group>
     </DimensionLayout>
