--- src/Vu/ui/UserManagermentDialog.java
@@ -1,628 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/AWTForms/Dialog.java to edit this template
- */
-package Vu.ui;
-
-/**
- *
- * @author Admin
- */
-public class UserManagermentDialog extends java.awt.Dialog {
-
-    /**
-     * Creates new form NewDialog
-     */
-    public UserManagermentDialog(java.awt.Frame parent, boolean modal) {
-        super(parent, modal);
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        findTextField1 = new AdminToolUtils.FindTextField();
-        toolbar1 = new AdminToolUtils.Toolbar();
-        jPanel3 = new javax.swing.JPanel();
-        jPanel1 = new javax.swing.JPanel();
-        jPanel2 = new javax.swing.JPanel();
-        findTextField2 = new AdminToolUtils.FindTextField();
-        jPanel4 = new javax.swing.JPanel();
-        pnlTitle = new javax.swing.JPanel();
-        lblNumber = new javax.swing.JLabel();
-        lblName = new javax.swing.JLabel();
-        lblUser = new javax.swing.JLabel();
-        lblStatus = new javax.swing.JLabel();
-        lblOption = new javax.swing.JLabel();
-        jSeparator2 = new javax.swing.JSeparator();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        jPanel5 = new javax.swing.JPanel();
-        jPanel6 = new javax.swing.JPanel();
-        lblNumber1 = new javax.swing.JLabel();
-        lblName1 = new javax.swing.JLabel();
-        lblUser1 = new javax.swing.JLabel();
-        btnOption1 = new javax.swing.JButton();
-        button1 = new AdminToolUtils.Button();
-        jPanel7 = new javax.swing.JPanel();
-        lblNumber2 = new javax.swing.JLabel();
-        lblName2 = new javax.swing.JLabel();
-        lblUser2 = new javax.swing.JLabel();
-        btnOption2 = new javax.swing.JButton();
-        button2 = new AdminToolUtils.Button();
-        jPanel15 = new javax.swing.JPanel();
-        lblNumber10 = new javax.swing.JLabel();
-        lblName10 = new javax.swing.JLabel();
-        lblUser10 = new javax.swing.JLabel();
-        btnOption10 = new javax.swing.JButton();
-        button3 = new AdminToolUtils.Button();
-        jPanel13 = new javax.swing.JPanel();
-        lblNumber8 = new javax.swing.JLabel();
-        lblName8 = new javax.swing.JLabel();
-        lblUser8 = new javax.swing.JLabel();
-        btnOption8 = new javax.swing.JButton();
-        button4 = new AdminToolUtils.Button();
-        jPanel16 = new javax.swing.JPanel();
-        lblNumber11 = new javax.swing.JLabel();
-        lblName11 = new javax.swing.JLabel();
-        lblUser11 = new javax.swing.JLabel();
-        btnOption11 = new javax.swing.JButton();
-        button5 = new AdminToolUtils.Button();
-
-        findTextField1.setText("findTextField1");
-
-        addWindowListener(new java.awt.event.WindowAdapter() {
-            public void windowClosing(java.awt.event.WindowEvent evt) {
-                closeDialog(evt);
-            }
-        });
-
-        jPanel3.setLayout(new java.awt.CardLayout());
-
-        jPanel1.setBackground(new java.awt.Color(36, 36, 36));
-
-        jPanel2.setBackground(new java.awt.Color(36, 36, 36));
-
-        findTextField2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-
-        javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
-        jPanel2.setLayout(jPanel2Layout);
-        jPanel2Layout.setHorizontalGroup(
-            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel2Layout.createSequentialGroup()
-                .addGap(79, 79, 79)
-                .addComponent(findTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 450, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        jPanel2Layout.setVerticalGroup(
-            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
-                .addContainerGap(77, Short.MAX_VALUE)
-                .addComponent(findTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(46, 46, 46))
-        );
-
-        jPanel4.setBackground(new java.awt.Color(36, 36, 36));
-
-        pnlTitle.setBackground(new java.awt.Color(36, 36, 36));
-        pnlTitle.setPreferredSize(new java.awt.Dimension(1073, 52));
-
-        lblNumber.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblNumber.setForeground(new java.awt.Color(199, 199, 199));
-        lblNumber.setText("STT");
-
-        lblName.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblName.setForeground(new java.awt.Color(199, 199, 199));
-        lblName.setText("Tên người dùng");
-
-        lblUser.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser.setForeground(new java.awt.Color(199, 199, 199));
-        lblUser.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser.setText("Tên tài khoản");
-
-        lblStatus.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblStatus.setForeground(new java.awt.Color(199, 199, 199));
-        lblStatus.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblStatus.setText("Tùy chọn");
-
-        lblOption.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblOption.setForeground(new java.awt.Color(199, 199, 199));
-        lblOption.setText("Trạng thái");
-
-        javax.swing.GroupLayout pnlTitleLayout = new javax.swing.GroupLayout(pnlTitle);
-        pnlTitle.setLayout(pnlTitleLayout);
-        pnlTitleLayout.setHorizontalGroup(
-            pnlTitleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlTitleLayout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addComponent(lblNumber, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(50, 50, 50)
-                .addComponent(lblName, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(52, 52, 52)
-                .addComponent(lblUser, javax.swing.GroupLayout.PREFERRED_SIZE, 294, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(81, 81, 81)
-                .addComponent(lblOption, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(lblStatus, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-        pnlTitleLayout.setVerticalGroup(
-            pnlTitleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlTitleLayout.createSequentialGroup()
-                .addGap(16, 16, 16)
-                .addGroup(pnlTitleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(lblNumber)
-                    .addComponent(lblName)
-                    .addComponent(lblUser)
-                    .addComponent(lblStatus)
-                    .addComponent(lblOption))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jSeparator2.setBackground(new java.awt.Color(199, 199, 199));
-        jSeparator2.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
-
-        jScrollPane1.setBackground(new java.awt.Color(0, 0, 0));
-        jScrollPane1.setBorder(null);
-
-        jPanel5.setBackground(new java.awt.Color(36, 36, 36));
-
-        jPanel6.setBackground(new java.awt.Color(36, 36, 36));
-
-        lblNumber1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblNumber1.setForeground(new java.awt.Color(255, 255, 255));
-        lblNumber1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblNumber1.setText("1");
-
-        lblName1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblName1.setForeground(new java.awt.Color(255, 255, 255));
-        lblName1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
-        lblName1.setText("Phan Quí Đức");
-
-        lblUser1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser1.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser1.setText("ducpq123");
-
-        btnOption1.setBackground(new java.awt.Color(36, 36, 36));
-        btnOption1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        btnOption1.setForeground(new java.awt.Color(255, 255, 255));
-        btnOption1.setText("Xóa");
-        btnOption1.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
-
-        button1.setForeground(new java.awt.Color(0, 0, 0));
-        button1.setText("Hiệu lực");
-        button1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-
-        javax.swing.GroupLayout jPanel6Layout = new javax.swing.GroupLayout(jPanel6);
-        jPanel6.setLayout(jPanel6Layout);
-        jPanel6Layout.setHorizontalGroup(
-            jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel6Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lblNumber1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(55, 55, 55)
-                .addComponent(lblName1, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(50, 50, 50)
-                .addComponent(lblUser1, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(65, 65, 65)
-                .addComponent(button1, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 123, Short.MAX_VALUE)
-                .addComponent(btnOption1, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(17, 17, 17))
-        );
-        jPanel6Layout.setVerticalGroup(
-            jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel6Layout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(lblNumber1)
-                    .addComponent(lblName1)
-                    .addComponent(lblUser1)
-                    .addComponent(btnOption1, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(button1, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(25, Short.MAX_VALUE))
-        );
-
-        jPanel7.setBackground(new java.awt.Color(36, 36, 36));
-
-        lblNumber2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblNumber2.setForeground(new java.awt.Color(255, 255, 255));
-        lblNumber2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblNumber2.setText("1");
-
-        lblName2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblName2.setForeground(new java.awt.Color(255, 255, 255));
-        lblName2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
-        lblName2.setText("Phan Quí Đức");
-
-        lblUser2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser2.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser2.setText("ducpq123");
-
-        btnOption2.setBackground(new java.awt.Color(36, 36, 36));
-        btnOption2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        btnOption2.setForeground(new java.awt.Color(255, 255, 255));
-        btnOption2.setText("Xóa");
-        btnOption2.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
-
-        button2.setForeground(new java.awt.Color(0, 0, 0));
-        button2.setText("Hiệu lực");
-        button2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-
-        javax.swing.GroupLayout jPanel7Layout = new javax.swing.GroupLayout(jPanel7);
-        jPanel7.setLayout(jPanel7Layout);
-        jPanel7Layout.setHorizontalGroup(
-            jPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel7Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lblNumber2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(55, 55, 55)
-                .addComponent(lblName2, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(50, 50, 50)
-                .addComponent(lblUser2, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(65, 65, 65)
-                .addComponent(button2, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(btnOption2, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(17, 17, 17))
-        );
-        jPanel7Layout.setVerticalGroup(
-            jPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel7Layout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addGroup(jPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(lblNumber2)
-                    .addComponent(lblName2)
-                    .addComponent(lblUser2)
-                    .addComponent(btnOption2, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(button2, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(21, Short.MAX_VALUE))
-        );
-
-        jPanel15.setBackground(new java.awt.Color(36, 36, 36));
-
-        lblNumber10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblNumber10.setForeground(new java.awt.Color(255, 255, 255));
-        lblNumber10.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblNumber10.setText("1");
-
-        lblName10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblName10.setForeground(new java.awt.Color(255, 255, 255));
-        lblName10.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
-        lblName10.setText("Phan Quí Đức");
-
-        lblUser10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser10.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser10.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser10.setText("ducpq123");
-
-        btnOption10.setBackground(new java.awt.Color(36, 36, 36));
-        btnOption10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        btnOption10.setForeground(new java.awt.Color(255, 255, 255));
-        btnOption10.setText("Xóa");
-        btnOption10.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
-
-        button3.setForeground(new java.awt.Color(0, 0, 0));
-        button3.setText("Hiệu lực");
-        button3.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-
-        javax.swing.GroupLayout jPanel15Layout = new javax.swing.GroupLayout(jPanel15);
-        jPanel15.setLayout(jPanel15Layout);
-        jPanel15Layout.setHorizontalGroup(
-            jPanel15Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel15Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lblNumber10, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(55, 55, 55)
-                .addComponent(lblName10, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(50, 50, 50)
-                .addComponent(lblUser10, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(65, 65, 65)
-                .addComponent(button3, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(btnOption10, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(17, 17, 17))
-        );
-        jPanel15Layout.setVerticalGroup(
-            jPanel15Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel15Layout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addGroup(jPanel15Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(lblNumber10)
-                    .addComponent(lblName10)
-                    .addComponent(lblUser10)
-                    .addComponent(btnOption10, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(button3, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(21, Short.MAX_VALUE))
-        );
-
-        jPanel13.setBackground(new java.awt.Color(36, 36, 36));
-
-        lblNumber8.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblNumber8.setForeground(new java.awt.Color(255, 255, 255));
-        lblNumber8.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblNumber8.setText("1");
-
-        lblName8.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblName8.setForeground(new java.awt.Color(255, 255, 255));
-        lblName8.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
-        lblName8.setText("Phan Quí Đức");
-
-        lblUser8.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser8.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser8.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser8.setText("ducpq123");
-
-        btnOption8.setBackground(new java.awt.Color(36, 36, 36));
-        btnOption8.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        btnOption8.setForeground(new java.awt.Color(255, 255, 255));
-        btnOption8.setText("Xóa");
-        btnOption8.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
-
-        button4.setBackground(new java.awt.Color(199, 199, 199));
-        button4.setText("Vô hiệu hóa");
-        button4.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-
-        javax.swing.GroupLayout jPanel13Layout = new javax.swing.GroupLayout(jPanel13);
-        jPanel13.setLayout(jPanel13Layout);
-        jPanel13Layout.setHorizontalGroup(
-            jPanel13Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel13Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lblNumber8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(55, 55, 55)
-                .addComponent(lblName8, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(50, 50, 50)
-                .addComponent(lblUser8, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(65, 65, 65)
-                .addComponent(button4, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(btnOption8, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(17, 17, 17))
-        );
-        jPanel13Layout.setVerticalGroup(
-            jPanel13Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel13Layout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addGroup(jPanel13Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(lblNumber8)
-                    .addComponent(lblName8)
-                    .addComponent(lblUser8)
-                    .addComponent(btnOption8, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(button4, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(21, Short.MAX_VALUE))
-        );
-
-        jPanel16.setBackground(new java.awt.Color(36, 36, 36));
-
-        lblNumber11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblNumber11.setForeground(new java.awt.Color(255, 255, 255));
-        lblNumber11.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblNumber11.setText("1");
-
-        lblName11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblName11.setForeground(new java.awt.Color(255, 255, 255));
-        lblName11.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
-        lblName11.setText("Phan Quí Đức");
-
-        lblUser11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser11.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser11.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser11.setText("ducpq123");
-
-        btnOption11.setBackground(new java.awt.Color(36, 36, 36));
-        btnOption11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        btnOption11.setForeground(new java.awt.Color(255, 255, 255));
-        btnOption11.setText("Xóa");
-        btnOption11.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
-
-        button5.setForeground(new java.awt.Color(0, 0, 0));
-        button5.setText("Hiệu lực");
-        button5.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-
-        javax.swing.GroupLayout jPanel16Layout = new javax.swing.GroupLayout(jPanel16);
-        jPanel16.setLayout(jPanel16Layout);
-        jPanel16Layout.setHorizontalGroup(
-            jPanel16Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel16Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lblNumber11, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(55, 55, 55)
-                .addComponent(lblName11, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(50, 50, 50)
-                .addComponent(lblUser11, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(65, 65, 65)
-                .addComponent(button5, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(btnOption11, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(17, 17, 17))
-        );
-        jPanel16Layout.setVerticalGroup(
-            jPanel16Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel16Layout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addGroup(jPanel16Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(lblNumber11)
-                    .addComponent(lblName11)
-                    .addComponent(lblUser11)
-                    .addComponent(btnOption11, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(button5, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(21, Short.MAX_VALUE))
-        );
-
-        javax.swing.GroupLayout jPanel5Layout = new javax.swing.GroupLayout(jPanel5);
-        jPanel5.setLayout(jPanel5Layout);
-        jPanel5Layout.setHorizontalGroup(
-            jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel5Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanel6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jPanel7, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jPanel15, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jPanel13, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jPanel16, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanel5Layout.setVerticalGroup(
-            jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel5Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jPanel6, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanel7, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanel15, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(jPanel13, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanel16, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(407, Short.MAX_VALUE))
-        );
-
-        jScrollPane1.setViewportView(jPanel5);
-
-        javax.swing.GroupLayout jPanel4Layout = new javax.swing.GroupLayout(jPanel4);
-        jPanel4.setLayout(jPanel4Layout);
-        jPanel4Layout.setHorizontalGroup(
-            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel4Layout.createSequentialGroup()
-                .addGap(0, 0, 0)
-                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 1243, Short.MAX_VALUE)
-                    .addComponent(jSeparator2)
-                    .addComponent(pnlTitle, javax.swing.GroupLayout.DEFAULT_SIZE, 1243, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanel4Layout.setVerticalGroup(
-            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel4Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(pnlTitle, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jSeparator2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 529, Short.MAX_VALUE)
-                .addContainerGap())
-        );
-
-        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
-        jPanel1.setLayout(jPanel1Layout);
-        jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-        jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, Short.MAX_VALUE))
-        );
-
-        jPanel3.add(jPanel1, "card2");
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
-        this.setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addComponent(toolbar1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, 0)
-                .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(toolbar1, javax.swing.GroupLayout.DEFAULT_SIZE, 835, Short.MAX_VALUE)
-            .addComponent(jPanel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * Closes the dialog
-     */
-    private void closeDialog(java.awt.event.WindowEvent evt) {//GEN-FIRST:event_closeDialog
-        setVisible(false);
-        dispose();
-    }//GEN-LAST:event_closeDialog
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                UserManagermentDialog dialog = new UserManagermentDialog(new java.awt.Frame(), true);
-                dialog.addWindowListener(new java.awt.event.WindowAdapter() {
-                    public void windowClosing(java.awt.event.WindowEvent e) {
-                        System.exit(0);
-                    }
-                });
-                dialog.setVisible(true);
-            }
-        });
-    }
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btnOption1;
-    private javax.swing.JButton btnOption10;
-    private javax.swing.JButton btnOption11;
-    private javax.swing.JButton btnOption2;
-    private javax.swing.JButton btnOption8;
-    private AdminToolUtils.Button button1;
-    private AdminToolUtils.Button button2;
-    private AdminToolUtils.Button button3;
-    private AdminToolUtils.Button button4;
-    private AdminToolUtils.Button button5;
-    private AdminToolUtils.FindTextField findTextField1;
-    private AdminToolUtils.FindTextField findTextField2;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JPanel jPanel13;
-    private javax.swing.JPanel jPanel15;
-    private javax.swing.JPanel jPanel16;
-    private javax.swing.JPanel jPanel2;
-    private javax.swing.JPanel jPanel3;
-    private javax.swing.JPanel jPanel4;
-    private javax.swing.JPanel jPanel5;
-    private javax.swing.JPanel jPanel6;
-    private javax.swing.JPanel jPanel7;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JSeparator jSeparator2;
-    private javax.swing.JLabel lblName;
-    private javax.swing.JLabel lblName1;
-    private javax.swing.JLabel lblName10;
-    private javax.swing.JLabel lblName11;
-    private javax.swing.JLabel lblName2;
-    private javax.swing.JLabel lblName8;
-    private javax.swing.JLabel lblNumber;
-    private javax.swing.JLabel lblNumber1;
-    private javax.swing.JLabel lblNumber10;
-    private javax.swing.JLabel lblNumber11;
-    private javax.swing.JLabel lblNumber2;
-    private javax.swing.JLabel lblNumber8;
-    private javax.swing.JLabel lblOption;
-    private javax.swing.JLabel lblStatus;
-    private javax.swing.JLabel lblUser;
-    private javax.swing.JLabel lblUser1;
-    private javax.swing.JLabel lblUser10;
-    private javax.swing.JLabel lblUser11;
-    private javax.swing.JLabel lblUser2;
-    private javax.swing.JLabel lblUser8;
-    private javax.swing.JPanel pnlTitle;
-    private AdminToolUtils.Toolbar toolbar1;
-    // End of variables declaration//GEN-END:variables
-}
