+++ src/Vu/ui/UserManagermentPanel.java
@@ -1,6 +1,6 @@
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/AWTForms/Dialog.java to edit this template
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
  */
 package Vu.ui;
 
@@ -8,13 +8,12 @@ package Vu.ui;
  *
  * @author Admin
  */
-public class CensorSongDialog extends java.awt.Dialog {
+public class UserManagermentPanel extends javax.swing.JPanel {
 
     /**
-     * Creates new form CensorSongDialog
+     * Creates new form UserManagermentPanel
      */
-    public CensorSongDialog(java.awt.Frame parent, boolean modal) {
-        super(parent, modal);
+    public UserManagermentPanel() {
         initComponents();
     }
 
@@ -23,13 +22,13 @@ public class CensorSongDialog extends java.awt.Dialog {
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
      */
+    @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
         jPanel1 = new javax.swing.JPanel();
         jPanel2 = new javax.swing.JPanel();
         findTextField2 = new AdminToolUtils.FindTextField();
-        cboSong = new AdminToolUtils.ComboBox();
         jPanel4 = new javax.swing.JPanel();
         pnlTitle = new javax.swing.JPanel();
         lblNumber = new javax.swing.JLabel();
@@ -45,48 +44,37 @@ public class CensorSongDialog extends java.awt.Dialog {
         lblName1 = new javax.swing.JLabel();
         lblUser1 = new javax.swing.JLabel();
         btnOption1 = new javax.swing.JButton();
-        lblUser9 = new javax.swing.JLabel();
+        button1 = new AdminToolUtils.ButtonDelete();
         jPanel7 = new javax.swing.JPanel();
         lblNumber2 = new javax.swing.JLabel();
         lblName2 = new javax.swing.JLabel();
+        lblUser2 = new javax.swing.JLabel();
         btnOption2 = new javax.swing.JButton();
-        lblUser4 = new javax.swing.JLabel();
-        lblUser3 = new javax.swing.JLabel();
+        button2 = new AdminToolUtils.ButtonDelete();
         jPanel15 = new javax.swing.JPanel();
         lblNumber10 = new javax.swing.JLabel();
         lblName10 = new javax.swing.JLabel();
+        lblUser10 = new javax.swing.JLabel();
         btnOption10 = new javax.swing.JButton();
-        lblUser5 = new javax.swing.JLabel();
-        lblUser12 = new javax.swing.JLabel();
+        button3 = new AdminToolUtils.ButtonDelete();
         jPanel13 = new javax.swing.JPanel();
         lblNumber8 = new javax.swing.JLabel();
         lblName8 = new javax.swing.JLabel();
+        lblUser8 = new javax.swing.JLabel();
         btnOption8 = new javax.swing.JButton();
-        lblUser6 = new javax.swing.JLabel();
-        lblUser14 = new javax.swing.JLabel();
+        button4 = new AdminToolUtils.ButtonDelete();
         jPanel16 = new javax.swing.JPanel();
         lblNumber11 = new javax.swing.JLabel();
         lblName11 = new javax.swing.JLabel();
+        lblUser11 = new javax.swing.JLabel();
         btnOption11 = new javax.swing.JButton();
-        lblUser7 = new javax.swing.JLabel();
-        lblUser13 = new javax.swing.JLabel();
-        toolbar1 = new AdminToolUtils.Toolbar();
-
-        addWindowListener(new java.awt.event.WindowAdapter() {
-            public void windowClosing(java.awt.event.WindowEvent evt) {
-                closeDialog(evt);
-            }
-        });
+        button5 = new AdminToolUtils.ButtonDelete();
 
         jPanel1.setBackground(new java.awt.Color(36, 36, 36));
 
         jPanel2.setBackground(new java.awt.Color(36, 36, 36));
 
         findTextField2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-
-        cboSong.setForeground(new java.awt.Color(0, 0, 0));
-        cboSong.setModel(new javax.swing.DefaultComboBoxModel(new String[] { "Tất cả bài hát", "Bài hát chưa kiểm duyệt" }));
-        cboSong.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
 
         javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
         jPanel2.setLayout(jPanel2Layout);
@@ -95,17 +83,13 @@ public class CensorSongDialog extends java.awt.Dialog {
             .addGroup(jPanel2Layout.createSequentialGroup()
                 .addGap(79, 79, 79)
                 .addComponent(findTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 450, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(cboSong, javax.swing.GroupLayout.PREFERRED_SIZE, 400, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(20, 20, 20))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         jPanel2Layout.setVerticalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
                 .addContainerGap(77, Short.MAX_VALUE)
-                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(findTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(cboSong, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addComponent(findTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(46, 46, 46))
         );
 
@@ -125,7 +109,7 @@ public class CensorSongDialog extends java.awt.Dialog {
         lblUser.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblUser.setForeground(new java.awt.Color(199, 199, 199));
         lblUser.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser.setText("Nhạc sĩ");
+        lblUser.setText("Tên tài khoản");
 
         lblStatus.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblStatus.setForeground(new java.awt.Color(199, 199, 199));
@@ -134,8 +118,7 @@ public class CensorSongDialog extends java.awt.Dialog {
 
         lblOption.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblOption.setForeground(new java.awt.Color(199, 199, 199));
-        lblOption.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblOption.setText("Ca sĩ");
+        lblOption.setText("Trạng thái");
 
         javax.swing.GroupLayout pnlTitleLayout = new javax.swing.GroupLayout(pnlTitle);
         pnlTitle.setLayout(pnlTitleLayout);
@@ -147,10 +130,10 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGap(50, 50, 50)
                 .addComponent(lblName, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(52, 52, 52)
-                .addComponent(lblUser, javax.swing.GroupLayout.PREFERRED_SIZE, 275, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(lblOption, javax.swing.GroupLayout.PREFERRED_SIZE, 240, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(37, 37, 37)
+                .addComponent(lblUser, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(81, 81, 81)
+                .addComponent(lblOption, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 91, Short.MAX_VALUE)
                 .addComponent(lblStatus, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addContainerGap())
         );
@@ -184,13 +167,13 @@ public class CensorSongDialog extends java.awt.Dialog {
 
         lblName1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblName1.setForeground(new java.awt.Color(255, 255, 255));
-        lblName1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
-        lblName1.setText("Mây hồng đưa lối");
+        lblName1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
+        lblName1.setText("Phan Quí Đức");
 
         lblUser1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblUser1.setForeground(new java.awt.Color(255, 255, 255));
         lblUser1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser1.setText("Taylor Swift");
+        lblUser1.setText("ducpq123");
 
         btnOption1.setBackground(new java.awt.Color(36, 36, 36));
         btnOption1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
@@ -198,10 +181,9 @@ public class CensorSongDialog extends java.awt.Dialog {
         btnOption1.setText("Xóa");
         btnOption1.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
 
-        lblUser9.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser9.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser9.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser9.setText("Đạt Ka");
+        button1.setForeground(new java.awt.Color(0, 0, 0));
+        button1.setText("Hiệu lực");
+        button1.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
 
         javax.swing.GroupLayout jPanel6Layout = new javax.swing.GroupLayout(jPanel6);
         jPanel6.setLayout(jPanel6Layout);
@@ -213,9 +195,9 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGap(55, 55, 55)
                 .addComponent(lblName1, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(50, 50, 50)
-                .addComponent(lblUser1, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(47, 47, 47)
-                .addComponent(lblUser9, javax.swing.GroupLayout.PREFERRED_SIZE, 249, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(lblUser1, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(65, 65, 65)
+                .addComponent(button1, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(btnOption1, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(17, 17, 17))
@@ -227,10 +209,10 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(lblNumber1)
                     .addComponent(lblName1)
-                    .addComponent(lblUser1, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUser1)
                     .addComponent(btnOption1, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser9, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(21, Short.MAX_VALUE))
+                    .addComponent(button1, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(25, Short.MAX_VALUE))
         );
 
         jPanel7.setBackground(new java.awt.Color(36, 36, 36));
@@ -242,8 +224,13 @@ public class CensorSongDialog extends java.awt.Dialog {
 
         lblName2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblName2.setForeground(new java.awt.Color(255, 255, 255));
-        lblName2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
-        lblName2.setText("Mây hồng đưa lối");
+        lblName2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
+        lblName2.setText("Phan Quí Đức");
+
+        lblUser2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser2.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser2.setText("ducpq123");
 
         btnOption2.setBackground(new java.awt.Color(36, 36, 36));
         btnOption2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
@@ -251,15 +238,9 @@ public class CensorSongDialog extends java.awt.Dialog {
         btnOption2.setText("Xóa");
         btnOption2.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
 
-        lblUser4.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser4.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser4.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser4.setText("Taylor Swift");
-
-        lblUser3.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser3.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser3.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser3.setText("Đạt Ka");
+        button2.setForeground(new java.awt.Color(0, 0, 0));
+        button2.setText("Hiệu lực");
+        button2.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
 
         javax.swing.GroupLayout jPanel7Layout = new javax.swing.GroupLayout(jPanel7);
         jPanel7.setLayout(jPanel7Layout);
@@ -271,10 +252,10 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGap(55, 55, 55)
                 .addComponent(lblName2, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(50, 50, 50)
-                .addComponent(lblUser4, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(47, 47, 47)
-                .addComponent(lblUser3, javax.swing.GroupLayout.PREFERRED_SIZE, 249, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 49, Short.MAX_VALUE)
+                .addComponent(lblUser2, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(65, 65, 65)
+                .addComponent(button2, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(btnOption2, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(17, 17, 17))
         );
@@ -285,9 +266,9 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGroup(jPanel7Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(lblNumber2)
                     .addComponent(lblName2)
+                    .addComponent(lblUser2)
                     .addComponent(btnOption2, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser4, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser3, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(button2, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addContainerGap(21, Short.MAX_VALUE))
         );
 
@@ -300,8 +281,13 @@ public class CensorSongDialog extends java.awt.Dialog {
 
         lblName10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblName10.setForeground(new java.awt.Color(255, 255, 255));
-        lblName10.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
-        lblName10.setText("Mây hồng đưa lối");
+        lblName10.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
+        lblName10.setText("Phan Quí Đức");
+
+        lblUser10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser10.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser10.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser10.setText("ducpq123");
 
         btnOption10.setBackground(new java.awt.Color(36, 36, 36));
         btnOption10.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
@@ -309,15 +295,9 @@ public class CensorSongDialog extends java.awt.Dialog {
         btnOption10.setText("Xóa");
         btnOption10.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
 
-        lblUser5.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser5.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser5.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser5.setText("Taylor Swift");
-
-        lblUser12.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser12.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser12.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser12.setText("Đạt Ka");
+        button3.setForeground(new java.awt.Color(0, 0, 0));
+        button3.setText("Hiệu lực");
+        button3.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
 
         javax.swing.GroupLayout jPanel15Layout = new javax.swing.GroupLayout(jPanel15);
         jPanel15.setLayout(jPanel15Layout);
@@ -329,10 +309,10 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGap(55, 55, 55)
                 .addComponent(lblName10, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(50, 50, 50)
-                .addComponent(lblUser5, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(47, 47, 47)
-                .addComponent(lblUser12, javax.swing.GroupLayout.PREFERRED_SIZE, 249, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(lblUser10, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(65, 65, 65)
+                .addComponent(button3, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 124, Short.MAX_VALUE)
                 .addComponent(btnOption10, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(17, 17, 17))
         );
@@ -343,9 +323,9 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGroup(jPanel15Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(lblNumber10)
                     .addComponent(lblName10)
+                    .addComponent(lblUser10)
                     .addComponent(btnOption10, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser5, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser12, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(button3, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addContainerGap(21, Short.MAX_VALUE))
         );
 
@@ -358,8 +338,13 @@ public class CensorSongDialog extends java.awt.Dialog {
 
         lblName8.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblName8.setForeground(new java.awt.Color(255, 255, 255));
-        lblName8.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
-        lblName8.setText("Mây hồng đưa lối");
+        lblName8.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
+        lblName8.setText("Phan Quí Đức");
+
+        lblUser8.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser8.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser8.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser8.setText("ducpq123");
 
         btnOption8.setBackground(new java.awt.Color(36, 36, 36));
         btnOption8.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
@@ -367,15 +352,9 @@ public class CensorSongDialog extends java.awt.Dialog {
         btnOption8.setText("Xóa");
         btnOption8.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
 
-        lblUser6.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser6.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser6.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser6.setText("Taylor Swift");
-
-        lblUser14.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser14.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser14.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser14.setText("Đạt Ka");
+        button4.setBackground(new java.awt.Color(199, 199, 199));
+        button4.setText("Vô hiệu hóa");
+        button4.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
 
         javax.swing.GroupLayout jPanel13Layout = new javax.swing.GroupLayout(jPanel13);
         jPanel13.setLayout(jPanel13Layout);
@@ -386,11 +365,11 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addComponent(lblNumber8, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(55, 55, 55)
                 .addComponent(lblName8, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(lblUser6, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(47, 47, 47)
-                .addComponent(lblUser14, javax.swing.GroupLayout.PREFERRED_SIZE, 249, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(47, 47, 47)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser8, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(65, 65, 65)
+                .addComponent(button4, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 125, Short.MAX_VALUE)
                 .addComponent(btnOption8, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(17, 17, 17))
         );
@@ -401,9 +380,9 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addGroup(jPanel13Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(lblNumber8)
                     .addComponent(lblName8)
+                    .addComponent(lblUser8)
                     .addComponent(btnOption8, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser6, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser14, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(button4, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addContainerGap(21, Short.MAX_VALUE))
         );
 
@@ -416,8 +395,13 @@ public class CensorSongDialog extends java.awt.Dialog {
 
         lblName11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
         lblName11.setForeground(new java.awt.Color(255, 255, 255));
-        lblName11.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/song.png"))); // NOI18N
-        lblName11.setText("Mây hồng đưa lối");
+        lblName11.setIcon(new javax.swing.ImageIcon(getClass().getResource("/AdminToolUtils/avatar.png"))); // NOI18N
+        lblName11.setText("Phan Quí Đức");
+
+        lblUser11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
+        lblUser11.setForeground(new java.awt.Color(255, 255, 255));
+        lblUser11.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblUser11.setText("ducpq123");
 
         btnOption11.setBackground(new java.awt.Color(36, 36, 36));
         btnOption11.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
@@ -425,15 +409,9 @@ public class CensorSongDialog extends java.awt.Dialog {
         btnOption11.setText("Xóa");
         btnOption11.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(36, 36, 36)));
 
-        lblUser7.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser7.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser7.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser7.setText("Taylor Swift");
-
-        lblUser13.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
-        lblUser13.setForeground(new java.awt.Color(255, 255, 255));
-        lblUser13.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblUser13.setText("Đạt Ka");
+        button5.setForeground(new java.awt.Color(0, 0, 0));
+        button5.setText("Hiệu lực");
+        button5.setFont(new java.awt.Font("Segoe UI", 0, 22)); // NOI18N
 
         javax.swing.GroupLayout jPanel16Layout = new javax.swing.GroupLayout(jPanel16);
         jPanel16.setLayout(jPanel16Layout);
@@ -444,24 +422,24 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addComponent(lblNumber11, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(55, 55, 55)
                 .addComponent(lblName11, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(52, 52, 52)
-                .addComponent(lblUser7, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 52, Short.MAX_VALUE)
-                .addComponent(lblUser13, javax.swing.GroupLayout.PREFERRED_SIZE, 249, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(42, 42, 42)
+                .addGap(50, 50, 50)
+                .addComponent(lblUser11, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(65, 65, 65)
+                .addComponent(button5, javax.swing.GroupLayout.PREFERRED_SIZE, 157, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(btnOption11, javax.swing.GroupLayout.PREFERRED_SIZE, 103, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(17, 17, 17))
         );
         jPanel16Layout.setVerticalGroup(
             jPanel16Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel16Layout.createSequentialGroup()
-                .addGap(17, 17, 17)
+                .addGap(20, 20, 20)
                 .addGroup(jPanel16Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(lblNumber11)
                     .addComponent(lblName11)
+                    .addComponent(lblUser11)
                     .addComponent(btnOption11, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser7, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUser13, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(button5, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addContainerGap(21, Short.MAX_VALUE))
         );
 
@@ -472,11 +450,14 @@ public class CensorSongDialog extends java.awt.Dialog {
             .addGroup(jPanel5Layout.createSequentialGroup()
                 .addContainerGap()
                 .addGroup(jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanel6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jPanel7, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jPanel15, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addComponent(jPanel13, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jPanel16, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addComponent(jPanel16, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addGroup(jPanel5Layout.createSequentialGroup()
+                        .addGroup(jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                            .addComponent(jPanel15, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(jPanel7, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(jPanel6, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                        .addGap(0, 0, Short.MAX_VALUE)))
                 .addContainerGap())
         );
         jPanel5Layout.setVerticalGroup(
@@ -488,11 +469,11 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addComponent(jPanel7, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jPanel15, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jPanel13, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(12, 12, 12)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jPanel16, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(413, Short.MAX_VALUE))
+                .addContainerGap(407, Short.MAX_VALUE))
         );
 
         jScrollPane1.setViewportView(jPanel5);
@@ -504,10 +485,11 @@ public class CensorSongDialog extends java.awt.Dialog {
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel4Layout.createSequentialGroup()
                 .addGap(0, 0, 0)
                 .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 1243, Short.MAX_VALUE)
+                    .addComponent(jScrollPane1)
                     .addComponent(jSeparator2)
-                    .addComponent(pnlTitle, javax.swing.GroupLayout.DEFAULT_SIZE, 1243, Short.MAX_VALUE))
-                .addContainerGap())
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel4Layout.createSequentialGroup()
+                        .addComponent(pnlTitle, javax.swing.GroupLayout.PREFERRED_SIZE, 1200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 0, Short.MAX_VALUE))))
         );
         jPanel4Layout.setVerticalGroup(
             jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -527,9 +509,9 @@ public class CensorSongDialog extends java.awt.Dialog {
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addContainerGap(10, Short.MAX_VALUE)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
+                .addGap(55, 55, 55))
         );
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -537,51 +519,20 @@ public class CensorSongDialog extends java.awt.Dialog {
                 .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 16, Short.MAX_VALUE))
+                .addGap(0, 14, Short.MAX_VALUE))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
         this.setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addComponent(toolbar1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, Short.MAX_VALUE)
-                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 1220, javax.swing.GroupLayout.PREFERRED_SIZE)
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addComponent(toolbar1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
         );
-
-        pack();
     }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * Closes the dialog
-     */
-    private void closeDialog(java.awt.event.WindowEvent evt) {//GEN-FIRST:event_closeDialog
-        setVisible(false);
-        dispose();
-    }//GEN-LAST:event_closeDialog
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                CensorSongDialog dialog = new CensorSongDialog(new java.awt.Frame(), true);
-                dialog.addWindowListener(new java.awt.event.WindowAdapter() {
-                    public void windowClosing(java.awt.event.WindowEvent e) {
-                        System.exit(0);
-                    }
-                });
-                dialog.setVisible(true);
-            }
-        });
-    }
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
@@ -590,7 +541,11 @@ public class CensorSongDialog extends java.awt.Dialog {
     private javax.swing.JButton btnOption11;
     private javax.swing.JButton btnOption2;
     private javax.swing.JButton btnOption8;
-    private AdminToolUtils.ComboBox cboSong;
+    private AdminToolUtils.ButtonDelete button1;
+    private AdminToolUtils.ButtonDelete button2;
+    private AdminToolUtils.ButtonDelete button3;
+    private AdminToolUtils.ButtonDelete button4;
+    private AdminToolUtils.ButtonDelete button5;
     private AdminToolUtils.FindTextField findTextField2;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JPanel jPanel13;
@@ -619,16 +574,10 @@ public class CensorSongDialog extends java.awt.Dialog {
     private javax.swing.JLabel lblStatus;
     private javax.swing.JLabel lblUser;
     private javax.swing.JLabel lblUser1;
-    private javax.swing.JLabel lblUser12;
-    private javax.swing.JLabel lblUser13;
-    private javax.swing.JLabel lblUser14;
-    private javax.swing.JLabel lblUser3;
-    private javax.swing.JLabel lblUser4;
-    private javax.swing.JLabel lblUser5;
-    private javax.swing.JLabel lblUser6;
-    private javax.swing.JLabel lblUser7;
-    private javax.swing.JLabel lblUser9;
+    private javax.swing.JLabel lblUser10;
+    private javax.swing.JLabel lblUser11;
+    private javax.swing.JLabel lblUser2;
+    private javax.swing.JLabel lblUser8;
     private javax.swing.JPanel pnlTitle;
-    private AdminToolUtils.Toolbar toolbar1;
     // End of variables declaration//GEN-END:variables
 }
