+++ src/swing/swavecomponent/AppItem.form
@@ -3,13 +3,13 @@
 <Form version="1.8" maxVersion="1.9" type="org.netbeans.modules.form.forminfo.JPanelFormInfo">
   <Properties>
     <Property name="maximumSize" type="java.awt.Dimension" editor="org.netbeans.beaninfo.editors.DimensionEditor">
-      <Dimension value="[249, 35]"/>
+      <Dimension value="[249, 45]"/>
     </Property>
     <Property name="minimumSize" type="java.awt.Dimension" editor="org.netbeans.beaninfo.editors.DimensionEditor">
-      <Dimension value="[249, 35]"/>
+      <Dimension value="[249, 45]"/>
     </Property>
     <Property name="preferredSize" type="java.awt.Dimension" editor="org.netbeans.beaninfo.editors.DimensionEditor">
-      <Dimension value="[249, 35]"/>
+      <Dimension value="[249, 45]"/>
     </Property>
   </Properties>
   <AuxValues>
@@ -28,6 +28,7 @@
   <Layout class="org.netbeans.modules.form.compat2.layouts.DesignFlowLayout">
     <Property name="alignment" type="int" value="3"/>
     <Property name="horizontalGap" type="int" value="20"/>
+    <Property name="verticalGap" type="int" value="10"/>
   </Layout>
   <SubComponents>
     <Component class="javax.swing.JLabel" name="lblIcon">
